This is hacky but there is no other way to shut up the compiler.
[git-p4: depot-paths = "//depot/qt/main/": change = 37044]

+++ src/kernel/qthread_p.h
@@ -323,7 +323,7 @@ public:
 #ifdef QT_CHECK_RANGE
 	    qWarning("Mutex unlocked from different thread than locker");
 	    qWarning("               was locked by %d, unlock attempt from %d",
-		     owner, QThread::currentThread());
+		     (int)owner, (int)QThread::currentThread());
 	    mutex_unlock(&mutex2);
 #endif
